Struct massa_final_state::final_state::FinalState

pub struct FinalState {
    pub(crate) config: FinalStateConfig,
    pub ledger: Box<dyn LedgerController>,
    pub async_pool: AsyncPool,
    pub pos_state: PoSFinalState,
    pub executed_ops: ExecutedOps,
    pub executed_denunciations: ExecutedDenunciations,
    pub mip_store: MipStore,
    pub last_start_period: u64,
    pub last_slot_before_downtime: Option<Slot>,
    pub db: ShareableMassaDBController,
}

Represents a final state (ledger, async pool, executed_ops, executed_de and the state of the PoS)

Fields
config: FinalStateConfig

execution state configuration

ledger: Box<dyn LedgerController>

final ledger associating addresses to their balance, executable bytecode and data

async_pool: AsyncPool

asynchronous pool containing messages sorted by priority and their data

pos_state: PoSFinalState

proof of stake state containing cycle history and deferred credits

executed_ops: ExecutedOps

executed operations

executed_denunciations: ExecutedDenunciations

executed denunciations

mip_store: MipStore

MIP store

last_start_period: u64

last_start_period

If start new network: set to 0

If from snapshot: retrieve from args

If from bootstrap: set during bootstrap

last_slot_before_downtime: Option<Slot>

last_slot_before_downtime

None if start new network

If from snapshot: retrieve from the slot attached to the snapshot

If from bootstrap: set during bootstrap

db: ShareableMassaDBController

the RocksDB instance used to write every final_state struct on disk

Implementations

impl FinalState

pub fn new( db: ShareableMassaDBController, config: FinalStateConfig, ledger: Box<dyn LedgerController>, selector: Box<dyn SelectorController>, mip_store: MipStore, reset_final_state: bool, ) -> Result<Self, FinalStateError>

Initializes a new FinalState

Arguments

config: the configuration of the final state to use for initialization

ledger: the instance of the ledger on disk. Used to apply changes to the ledger.

selector: the pos selector. Used to send draw inputs when a new cycle is completed.

reset_final_state: if true, we only keep the ledger, and we reset the other fields of the final state

fn interpolate_downtime(&mut self) -> Result<(), FinalStateError>

Once we created a FinalState from a snapshot, we need to edit it to attach at the end_slot and handle the downtime. This basically recreates the history of the final_state, without executing the slots.

fn interpolate_single_cycle( &mut self, current_slot: Slot, end_slot: Slot, ) -> Result<(), FinalStateError>

This helper function is to be called if the downtime does not span over multiple cycles

fn interpolate_multiple_cycles( &mut self, current_slot: Slot, end_slot: Slot, current_slot_cycle: u64, end_slot_cycle: u64, ) -> Result<(), FinalStateError>

This helper function is to be called if the downtime spans over multiple cycles

fn feed_cycle_hash_and_selector_for_interpolation( &mut self, cycle: u64, ) -> Result<(), FinalStateError>

Used during interpolation, when a new cycle is set as completed

fn _finalize(&mut self, slot: Slot, changes: StateChanges) -> AnyResult<()>

fn _is_db_valid(&self) -> AnyResult<()>

Internal function called by is_db_valid

pub fn new_derived_from_snapshot( db: ShareableMassaDBController, config: FinalStateConfig, ledger: Box<dyn LedgerController>, selector: Box<dyn SelectorController>, mip_store: MipStore, last_start_period: u64, ) -> Result<Self, FinalStateError>

Initializes a FinalState from a snapshot.

Arguments

db: A type that implements the MassaDBController trait. Used to read and write to the database.

config: the configuration of the final state to use for initialization

ledger: the instance of the ledger on disk. Used to apply changes to the ledger.

selector: the pos selector. Used to send draw inputs when a new cycle is completed.

last_start_period: at what period we should attach the final_state